+++ package.json
@@ -55,7 +55,8 @@
     "@angular/language-service": "^9.0.0",
     "@babel/core": "^7.8.4",
     "@pickra/copy-code-block": "^1.2.0",
-    "@storybook/html": "^5.3.12",
+    "@storybook/addon-storysource": "next",
+    "@storybook/html": "next",
     "@types/jasmine": "~3.5.3",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "^13.7.0",
